refactor: break down large components into smaller chunks (#52)
