+++ src/components/VisualHelper.vue
@@ -1,0 +1,10 @@
+<script setup lang="ts">
+</script>
+
+<template>
+  <TresAxesHelper />
+  <TresGridHelper />
+</template>
+
+<style scoped>
+</style>
