+++ src/views/GameView.vue
@@ -2,12 +2,11 @@
 import { NoToneMapping, SRGBColorSpace, VSMShadowMap } from 'three'
 import { TresCanvas } from '@tresjs/core'
 import type { TresCanvasProps } from '@tresjs/core/dist/src/components/TresCanvas.vue.js'
-import Game from '@/components/Game.vue'
-import useRegisteredForSelectingModelStore from '@/composables/useRegisteredForSelectingModelStore'
+import GameEngine from '@/components/GameEngine.vue'
 import useClickedModelNodeStore from '@/composables/useSelectedModelsStore'
+import ClickedModelDebugUI from '@/components/ClickedModelDebugUI.vue'
 
 const clickedModelNodeStore = useClickedModelNodeStore()
-const registeredForSelectingModelStore = useRegisteredForSelectingModelStore()
 
 const gl: TresCanvasProps = {
   alpha: false,
@@ -27,32 +26,10 @@ const gl: TresCanvasProps = {
     v-bind="gl"
     @pointer-missed="() => clickedModelNodeStore.unselect()"
   >
-    <Game />
+    <GameEngine />
   </TresCanvas>
 
-  <div class="absolute top-0 left-0 z-20 flex flex-col gap-4">
-    <div class="bg-red-200 p-2">
-      registered:
-      <pre>{{ Array.from(registeredForSelectingModelStore.getRegisteredAll).map(e => `ID ${e}`) }}</pre>
-    </div>
-    <div class="bg-red-200 p-2">
-      selected:
-      <pre>{{ clickedModelNodeStore.getSelected?.id ?? '-' }}</pre>
-    </div>
-    <div class="bg-red-200 p-2">
-      move selected:<br>
-      <button
-        class="bg-blue-200 py-1 px-2 rounded-md"
-        @click="() => {
-          if (clickedModelNodeStore.hasSelected()){
-            clickedModelNodeStore.getSelected!.position.x += 0.1
-          }
-        }"
-      >
-        move
-      </button>
-    </div>
-  </div>
+  <ClickedModelDebugUI />
 </template>
 
 <style scoped>
